requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
astroid==4.0.1
attrs==25.4.0
black==25.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
clearml==2.0.2
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
datasets==4.4.1
dill==0.4.0
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
furl==2.1.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.1
idna==3.11
isort==7.0.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lightning-utilities==0.15.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mccabe==0.7.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.4
orderedmultidict==1.0.1
packaging==25.0
pandas==2.3.3
pathlib2==2.3.7.post1
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
propcache==0.4.1
psutil==7.1.3
pyarrow==22.0.0
PyJWT==2.10.1
pylint==4.0.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tomlkit==0.13.3
torch==2.9.0+cu130
torchmetrics==1.8.2
torchvision==0.24.0+cu130
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0